+++ src/app/controllers/all.js
@@ -15,4 +15,5 @@ define([
   './opentsdbTargetCtrl',
   './annotationsEditorCtrl',
   './templateEditorCtrl',
+  './jsonEditorCtrl',
 ], function () {});